feat(ga): add basic implementation with tests for EA
Added `EAComponents` to make working with optimizers a little bit
easier/handleable. `EA` is a struct responsible for optimizing
populations and storing experiment details (possibly add metadata?).
`EA` is constructed via `EABuilder` to ensure proper construction of the
method. Helper macro is `ea_components` is created, to decrease the
amount of boilerplate (though not sure if it's needed here, or if macros
should be gathered somewhere else).